sidecar/pyproject.toml
[project]
name = "kibrary-sidecar"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "kiutils>=1.4.8",
  "JLC2KiCadLib>=1.0.0",
  "httpx>=0.27.0",
  "pydantic>=2.9.0",
  "GitPython>=3.1.43"
]

[project.optional-dependencies]
dev = ["pytest>=8.3.0", "pytest-asyncio>=0.24.0", "pyinstaller>=6.10.0", "respx>=0.21"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["kibrary_sidecar"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
